#230: refactor 提取 LangSelector 公共组件消除重复代码
