+++ frontend/src/views/LoginView.vue
@@ -6,12 +6,7 @@
           <h1 class="login-title">{{ $t('app.title') }}</h1>
           <p class="login-subtitle">{{ $t('login.subtitle') }}</p>
         </div>
-        <div class="lang-selector">
-          <select v-model="currentLocale" class="lang-select">
-            <option value="zh-CN">中文</option>
-            <option value="en-US">English</option>
-          </select>
-        </div>
+        <LangSelector />
       </div>
 
       <form class="login-form" @submit.prevent="handleLogin">
@@ -64,15 +59,14 @@
 </template>
 
 <script setup lang="ts">
-import { ref, reactive, computed } from 'vue'
+import { ref, reactive } from 'vue'
 import { useRouter } from 'vue-router'
 import { useI18n } from 'vue-i18n'
 import { useUserStore } from '@/stores/user'
-import { setLocale } from '@/i18n'
-import type { Locale } from '@/i18n'
+import LangSelector from '@/components/common/LangSelector.vue'
 
 const router = useRouter()
-const { t, locale } = useI18n()
+const { t } = useI18n()
 const userStore = useUserStore()
 
 const form = reactive({
@@ -82,14 +76,6 @@ const form = reactive({
 
 const loading = ref(false)
 const error = ref('')
-
-// 当前语言
-const currentLocale = computed({
-  get: () => locale.value as Locale,
-  set: (value: Locale) => {
-    setLocale(value)
-  }
-})
 
   const handleLogin = async () => {
     error.value = ''
@@ -139,31 +125,6 @@ const currentLocale = computed({
 
 .header-content {
   flex: 1;
-}
-
-.lang-selector {
-  margin-left: 16px;
-}
-
-.lang-select {
-  padding: 6px 12px;
-  background: #f5f5f5;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  font-size: 13px;
-  color: #555;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.lang-select:hover {
-  border-color: #667eea;
-}
-
-.lang-select:focus {
-  outline: none;
-  border-color: #667eea;
-  box-shadow: 0 0 0 2px rgba(102, 126, 234, 0.1);
 }
 
 .login-title {
